Consolidate ad-hoc MIR lints into real pass-manager-based MIR lints
